requirements.txt
streamlit==1.31.1
pandas==2.0.3
fpdf2==2.7.11
requests==2.31.0
apscheduler==3.10.4
pywin32==306; sys_platform == 'win32'
python-dateutil==2.8.2
sqlite3
pytz